Refactor imports after moving things to App component folder
