+++ src/APICalls.js
@@ -1,5 +1,5 @@
-function getAPIData() {
-  return fetch ('http://localhost:3001/api/v1/reservations')
+function getAPIData(info) {
+  return fetch (`http://localhost:3001/api/v1/${info}`)
   .then((res)=> res.json())
 }
