+++ src/components/App/App.css
@@ -4,7 +4,7 @@
   display: flex;
   flex-direction: column;
   align-items: center;
-  background-image: url('../assets/cafe-img.jpeg');
+  background-image: url('/Users/cieragrace/turing_work/Mod3/turing_cafe/turing-cafe-ui/src/assets/cafe-img.jpeg');
   width: 100%;
   height: 100vh;
   background-repeat: no-repeat;
