+++ src/components/App/App.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import './App.css';
+import '../App/App.css'
 import getAPIData from '../../APICalls';
 import Reservations from '../Reservations/reservations'
 
@@ -12,8 +12,8 @@ class App extends Component {
   }
 
   componentDidMount = () => {
-    getAPIData()
-    .then((data)=> this.setState({ reservations: data}))
+    getAPIData('reservations')
+    .then((data)=> this.setState({ reservations: data.reservations}))
     .catch((error)=> console.log(error))
   }
 
@@ -26,7 +26,7 @@ class App extends Component {
       <div className="App">
         <h1 className='app-title'>Turing Cafe Reservations</h1>
         <div className='resy-form'></div>
-        <div className='resy-container'></div>
+        <Reservations reservations={this.state.reservations}/>
       </div>
     )
   }
