+++ src/components/Card/Card.js
@@ -1,13 +1,13 @@
 import './Card.css'
 import React from 'react'
 
-const Card = () => {
+const Card = ({name, date, time, number}) => {
   return (
     <div className='reservation-container'>
-      <h2 className='name'></h2>
-      <p className='info'></p>
-      <p className='info'></p>
-      <p className='info'></p>
+      <h2 className='name'>{name}</h2>
+      <p className='info'>{date}</p>
+      <p className='info'>{time}</p>
+      <p className='info'>{number}</p>
       <button className='cancel-btn'>Cancel</button>
     </div>
   )
